requirements.txt
Flask==2.3.2
mysql-connector-python==8.1.0
Pillow==9.4.0
Flask-Mail==0.9.1
